LC-add-to-array-form-of-integer [Runtime: N/A; Memory: N/A]

+++ math/add-to-array-form-of-integer.cpp
@@ -1,12 +1,12 @@
 class Solution {
 public:
     vector<int> addToArrayForm(vector<int>& num, int k) {
-        int n = 0;
+        long long n = 0;
         vector<int> res;
         for(int i = 0; i < num.size(); i++){
             n = n + num[i]*pow(10,num.size()-1-i);
         }
-        int ans = n + k;
+        long long ans = (long long)n + k;
         while(ans > 0){
             res.insert(res.begin(),ans%10);
             ans = ans/10;
